#4 add testing unit and integration. Done.
