+++ src/tests/unit/components/Button.test.tsx
@@ -1,4 +1,3 @@
-// Button.test.tsx
 import { render, screen } from "@testing-library/react";
 import user from "@testing-library/user-event";
 import "@testing-library/jest-dom";
